Update pipelines for builders without pem onboard
While updating our build images for these pipelines, we have removed
the "aether-qa.pem" file that was baked into the old images. This
will now be accessed from the Jenkins system instead.

Because the PEM path is expected to be the same during various steps
of the pipeline, using "withCredentials" in each step doesn't work,
due to the temp name changing with each invocation. Instead, we will
copy it to a location contained in an env var, as reference that env
var rather than having a hardcoded path.

Signed-off-by: Eric Ball <eball@linuxfoundation.org>