Update all pipelines for latest builders
A requirement for Ubuntu 24.04 and later versions to come is to have
all Python packages installed into a venv, rather than globally. In
order to keep the pipelines compatible with both older and newer
build images, the path to the venv is checked before each shell step
(except those utilizing only the aws-cli), and if the venv is present
on the agent, it is activated.

To improve security, the key used to make connections between agents
is no longer included on the image, but is put in place at run-time
from the Jenkins server.

Finally, there are some small changes, including adding the file name
to the top of each file (helpful when troubleshooting or replaying
pipelines) and cleaning up whitespace.

Signed-off-by: Eric Ball <eball@linuxfoundation.org>